requirements.txt
jsonlines
tqdm
matplotlib
ipykernel
pandas
python-dotenv
requests